move out common execute command function
